Devops/nightly build (#979)
NOTE: this will likely require a bit of tweaking after merge, only so
much you can test these outside of github. Basic concept of the nightly
was already established for OpenDCS and there are only minor differences
here.

Create build automation for:
 - Nightly builds
 - Tag builds

Also includes support for forcing the nightly to run. Such forcing was
not added to the Tag based build since that is triggered by creating a
tag anyways.